push

+++ philo_bonus/strtrim.c
@@ -6,7 +6,7 @@
 /*   By: ebennamr <ebennamr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/29 13:11:40 by ebennamr          #+#    #+#             */
-/*   Updated: 2023/05/06 12:46:27 by ebennamr         ###   ########.fr       */
+/*   Updated: 2023/05/06 21:00:00 by ebennamr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ char	*ft_substr(char *str, unsigned int start, size_t len)
 	i = 0;
 	if (!str)
 		return (NULL);
-	if (len > ft_strlen(str))
+	if ((int)len > ft_strlen(str))
 		len = ft_strlen(str);
 	strout = (char *)malloc(len * sizeof(char) + 1);
 	protec_error_pt(strout, ERR_OUT);
